Ask requested Justin wanted changes that I'd made for my personal uses. I put them here.
Fixed many problems including this one https://github.com/JustinSamaKun/ChestShops/issues/1
Also, added a configuration if any server owner has a different language for their server. With the configuration has a toggle able item display, if you found duplication bug with another mod, it was a hard fix for this one because the mod does take a higher priority, then plugins trying to override it. tried multiple ways of patching it but ultimately just disable it had the biggest effect on it.